norminete

+++ src/hooks.c
@@ -1,14 +1,26 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   hooks.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: dgorold- <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2019/03/18 14:24:08 by dgorold-          #+#    #+#             */
+/*   Updated: 2019/09/23 23:36:50 by dgorold-         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../includes/fractol.h"
 
-int 			deal_key(int key, t_fcl *f)
+int					deal_key(int key, t_fcl *f)
 {
-	if (key == 123) // left
+	if (key == 123)
 		f->offset_x += f->tmp1;
-	if (key == 124) // right
+	if (key == 124)
 		f->offset_x -= f->tmp1;
-	if (key == 126) // up
+	if (key == 126)
 		f->offset_y -= f->tmp1;
-	if (key == 125) // down
+	if (key == 125)
 		f->offset_y += f->tmp1;
 	if (key == 6)
 		f->stop_move = (f->stop_move == 1) ? 0 : 1;
@@ -18,30 +30,33 @@ int 			deal_key(int key, t_fcl *f)
 	return (0);
 }
 
-static int 			choose_zoom(t_fcl *f, int x, int y, int key)
+static int			choose_zoom(t_fcl *f, int x, int y, int key)
 {
-	t_complex	mouse;
+	t_complex		mouse;
 
 	if (key == 4)
 	{
-		mouse = init_complex((double)x / (WTH / (f->max.re - f->min.re)) + f->min.re,
-							 (double)y / (HGT / (f->max.im - f->min.im)) + f->min.im);
+		mouse = init_complex((double)x / (WTH /
+				(f->max.re - f->min.re)) + f->min.re,
+				(double)y / (HGT / (f->max.im - f->min.im)) + f->min.im);
 		apply_zoom(f, mouse, 1.25);
 	}
 	if (key == 5)
 	{
-		if (f->min.im == -2.0f && f->min.re == -2.0 && f->max.im == 2.0 && f->max.re == 2.0)
+		if (f->min.im == -2.0f && f->min.re == -2.0 &&
+		f->max.im == 2.0 && f->max.re == 2.0)
 			return (0);
-		mouse = init_complex((double)x / (WTH / (f->max.re - f->min.re)) + f->min.re,
-							 (double)y / (HGT / (f->max.im - f->min.im)) + f->min.im);
+		mouse = init_complex((double)x / (WTH /
+				(f->max.re - f->min.re)) + f->min.re,
+				(double)y / (HGT / (f->max.im - f->min.im)) + f->min.im);
 		apply_zoom(f, mouse, 0.8);
 	}
 	return (1);
 }
 
-int 			deal_mouse(int key, int x, int y, t_fcl *f)
+int					deal_mouse(int key, int x, int y, t_fcl *f)
 {
-	y = (int)HGT - y;
+	y = HGT - y;
 	if (key == 1)
 		f->max_i += 40;
 	if (key == 2)
@@ -53,10 +68,10 @@ int 			deal_mouse(int key, int x, int y, t_fcl *f)
 	return (0);
 }
 
-int 			mouse_move(int x_, int y_, t_fcl *f)
+int					mouse_move(int x_, int y_, t_fcl *f)
 {
-	double		x;
-	double 		y;
+	double			x;
+	double			y;
 
 	x = (double)x_;
 	y = (double)y_;
@@ -72,7 +87,7 @@ int 			mouse_move(int x_, int y_, t_fcl *f)
 	return (0);
 }
 
-void			hooks(t_fcl *f)
+void				hooks(t_fcl *f)
 {
 	mlx_hook(f->mlx.win, 2, 5, deal_key, f);
 	mlx_hook(f->mlx.win, 6, 0, mouse_move, f);
